// micro-apps/spinal-app-viewer-tickets/src/services/store/appDataStore/actions.ts
/*
 * Copyright 2022 SpinalCom - www.spinalcom.com
 *
 * This file is part of SpinalCore.
 *
 * Please read all of the following terms and conditions
 * of the Free Software license Agreement ("Agreement")
 * carefully.
 *
 * This Agreement is a legally binding contract between
 * the Licensee (as defined below) and SpinalCom that
 * sets forth the terms and conditions that govern your
 * use of the Program. By installing and/or using the
 * Program, you agree to abide by all the terms and
 * conditions stated or referenced herein.
 *
 * If you do not agree to abide by these terms and
 * conditions, do not demonstrate your acceptance and do
 * not install or use the Program.
 * You should have received a copy of the license along
 * with this file. If not, see
 * <http://resources.spinalcom.com/licenses.pdf>.
 */

import {
  getBuildings,
  getBuildingById,
} from "../../spinalAPI/GeographicContext/getBuildings";
import { IGetAllBuildingsRes } from "../../../interfaces/IGetAllBuildingsRes";
import { SpinalAPI } from "../../spinalAPI/SpinalAPI";
import { MutationTypes } from "./mutations";
import {
  equipment_read_details_multiple,
  getEquipments,
  getFloors,
  getRooms,
  getBuildingReferenceObecjts,
  postBIMObjectInfo,
  getStaticDetails,
  getStaticDetailsEquipement,
  getMultipleReferenceObjects,
  getBuilding,
} from "../../spinalAPI/GeographicContext/geographicContext";
import {
  addTicketDoc,
  createTicket,
  archiveTicket,
  moveToStepTicket,
  addNoteTicket,
  modify_ticket,
} from "../../spinalAPI/Workflow & ticket/ticketContext";
import type {
  IEquipmentItem,
  ISpaceSelectorItem,
  IZoneItem,
} from "../../../components/SpaceSelector";
import { INodeItem } from "../../../interfaces/INodeItem";
import {
  IViewInfoBody,
  IViewInfoItemRes,
} from "../../spinalAPI/GeographicContext/getViewInfo";
import {
  ActionTypes,
  ApiIteratorStoreRecordNumberType,
  ApiIteratorStoreRecordStringType,
  ApiIteratorStoreType,
  AugmentedActionContextAppData,
} from "../../../interfaces/vuexStoreTypes";
import {
  getGroupsItems,
  getAllCategoriesTree,
} from "../../spinalAPI/GeographicContext/groupsItems";
import {
  getWorkflowList,
  getProcessList,
  loadTickets,
  filterTicketsOnPosition,
} from "../../store/appDataStore/utils/ticketUtils";
import SpriteManager from "../../../../../../global-components/viewer/manager/spriteManager";
import ViewerManager from "../../../../../../global-components/viewer/manager/viewerManager";
import { IConfig } from "../../../interfaces/IConfig";
import { classifyItemByBimFileId } from "./utils/openViewer";

const ApiIteratorStore: ApiIteratorStoreType &
  ApiIteratorStoreRecordStringType &
  ApiIteratorStoreRecordNumberType = {};

export const actions = {
  async [ActionTypes.GET_BUILDINGS](
    { commit, state }: AugmentedActionContextAppData,
    { patrimoineId, forceUpdate }
  ): Promise<IGetAllBuildingsRes[]> {
    const spinalAPI = SpinalAPI.getInstance();
    if (typeof ApiIteratorStore[ActionTypes.GET_BUILDINGS] === "undefined") {
      ApiIteratorStore[ActionTypes.GET_BUILDINGS] = {};
    }

    if (
      typeof ApiIteratorStore[ActionTypes.GET_BUILDINGS][patrimoineId] ===
        "undefined" ||
      forceUpdate === true
    ) {
      ApiIteratorStore[ActionTypes.GET_BUILDINGS][patrimoineId] =
        spinalAPI.createIteratorCall(getBuildings, patrimoineId);
    }

    const buildings = await ApiIteratorStore[ActionTypes.GET_BUILDINGS][
      patrimoineId
    ]!.next();
    commit(MutationTypes.SET_BUILDINGS, buildings.value);
    return buildings.value;
  },

  async [ActionTypes.GET_BUILDING_BY_ID](
    { commit, state }: AugmentedActionContextAppData,
    { buildingId, forceUpdate }
  ): Promise<IGetAllBuildingsRes> {
    const spinalAPI = SpinalAPI.getInstance();
    if (
      typeof ApiIteratorStore[ActionTypes.GET_BUILDING_BY_ID] === "undefined"
    ) {
      ApiIteratorStore[ActionTypes.GET_BUILDING_BY_ID] = {};
    }

    if (
      typeof ApiIteratorStore[ActionTypes.GET_BUILDING_BY_ID][buildingId] ===
        "undefined" ||
      forceUpdate === true
    ) {
      ApiIteratorStore[ActionTypes.GET_BUILDING_BY_ID][buildingId] =
        spinalAPI.createIteratorCall(getBuildingById, buildingId);
    }

    const building = await ApiIteratorStore[ActionTypes.GET_BUILDING_BY_ID][
      buildingId
    ]!.next();
    return building.value;
  },
  async [ActionTypes.GET_BOS_BUILDING](
    { commit }: AugmentedActionContextAppData,
    { buildingId }: { buildingId: string }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await getBuilding(buildingId);
      return result;
    } catch (error) {
      console.error(
        "Erreur lors de la récupération des objets de référence:",
        error
      );
      throw error;
    }
  },

  async [ActionTypes.GET_BUILDING_REFERENCE_OBJECTS](
    { commit }: AugmentedActionContextAppData,
    { buildingId, patrimoineId, forceUpdate }
  ): Promise<IZoneItem[]> {
    const spinalAPI = SpinalAPI.getInstance();
    if (
      typeof ApiIteratorStore[ActionTypes.GET_BUILDING_REFERENCE_OBJECTS] ===
      "undefined"
    ) {
      ApiIteratorStore[ActionTypes.GET_BUILDING_REFERENCE_OBJECTS] = {};
    }

    const buildingRefObjs =
      ApiIteratorStore[ActionTypes.GET_BUILDING_REFERENCE_OBJECTS]!;
    if (
      typeof buildingRefObjs[buildingId] === "undefined" ||
      forceUpdate === true
    ) {
      buildingRefObjs[buildingId] = spinalAPI.createIteratorCall(
        getBuildingReferenceObecjts,
        patrimoineId,
        buildingId
      );
    }

    const RefObjs = await buildingRefObjs[buildingId].next();
    // commit(MutationTypes.SET_BUILDING_REF_OBJS, { id: buildingId, items: floors.value });
    return RefObjs.value;
  },

  async [ActionTypes.GET_FLOORS](
    { commit }: AugmentedActionContextAppData,
    { buildingId, patrimoineId, forceUpdate }
  ): Promise<IZoneItem[]> {
    const spinalAPI = SpinalAPI.getInstance();
    if (typeof ApiIteratorStore[ActionTypes.GET_FLOORS] === "undefined") {
      ApiIteratorStore[ActionTypes.GET_FLOORS] = {};
    }

    const floorObjStore = ApiIteratorStore[ActionTypes.GET_FLOORS]!;
    if (
      typeof floorObjStore[buildingId] === "undefined" ||
      forceUpdate === true
    ) {
      floorObjStore[buildingId] = spinalAPI.createIteratorCall(
        getFloors,
        patrimoineId,
        buildingId
      );
    }

    const floors = await floorObjStore[buildingId].next();
    commit(MutationTypes.SET_FLOORS, { id: buildingId, items: floors.value });
    return floors.value;
  },

  async [ActionTypes.GET_ROOMS](
    { commit }: AugmentedActionContextAppData,
    { buildingId, patrimoineId, floorId, id, forceUpdate }: any
  ): Promise<IZoneItem[]> {
    const spinalAPI = SpinalAPI.getInstance();
    if (typeof ApiIteratorStore[ActionTypes.GET_ROOMS] === "undefined") {
      ApiIteratorStore[ActionTypes.GET_ROOMS] = {};
    }
    const floorObjStore = ApiIteratorStore[ActionTypes.GET_ROOMS]!;

    if (typeof floorObjStore[id] === "undefined" || forceUpdate === true) {
      floorObjStore[id] = spinalAPI.createIteratorCall(
        getRooms,
        patrimoineId,
        buildingId,
        floorId,
        id
      );
    }
    const floors = await floorObjStore[id].next();
    commit(MutationTypes.SET_ROOMS, { id: id, items: floors.value });
    return floors.value;
  },

  async [ActionTypes.GET_EQUIPMENTS](
    { commit }: AugmentedActionContextAppData,
    { floorId, roomId, patrimoineId, buildingId, id, forceUpdate }: any
  ): Promise<IEquipmentItem[]> {
    const spinalAPI = SpinalAPI.getInstance();

    if (typeof ApiIteratorStore[ActionTypes.GET_EQUIPMENTS] === "undefined") {
      ApiIteratorStore[ActionTypes.GET_EQUIPMENTS] = {};
    }

    const roomObjStore = ApiIteratorStore[ActionTypes.GET_EQUIPMENTS]!;

    if (typeof roomObjStore[id] === "undefined" || forceUpdate === true) {
      roomObjStore[id] = spinalAPI.createIteratorCall(
        getEquipments,
        patrimoineId,
        buildingId,
        floorId,
        roomId,
        id
      );
    }

    const equipments = await roomObjStore[id].next();
    commit(MutationTypes.SET_EQUIPMENTS, { id: id, items: equipments.value });
    return equipments.value;
  },

  async [ActionTypes.GET_GROUPS_ITEMS](
    { commit, state }: AugmentedActionContextAppData,
    { forceUpdate, config, buildingId }
  ): Promise<{ [key: string]: INodeItem }> {
    const spinalAPI = SpinalAPI.getInstance();

    if (typeof ApiIteratorStore[ActionTypes.GET_GROUPS_ITEMS] === "undefined") {
      ApiIteratorStore[ActionTypes.GET_GROUPS_ITEMS] = {};
    }

    if (
      typeof ApiIteratorStore[ActionTypes.GET_GROUPS_ITEMS][buildingId] ===
        "undefined" ||
      forceUpdate
    ) {
      ApiIteratorStore[ActionTypes.GET_GROUPS_ITEMS][buildingId] =
        spinalAPI.createIteratorCall(getGroupsItems, config, buildingId);
    }

    const items = await ApiIteratorStore[ActionTypes.GET_GROUPS_ITEMS][
      buildingId
    ].next();
    return items?.value;
  },

  async [ActionTypes.GET_CATEGORIES_TREE](
    { commit, state }: AugmentedActionContextAppData,
    { forceUpdate, buildingId, context }
  ): Promise<{ [key: string]: INodeItem }> {
    const spinalAPI = SpinalAPI.getInstance();

    if (
      typeof ApiIteratorStore[ActionTypes.GET_CATEGORIES_TREE] === "undefined"
    ) {
      ApiIteratorStore[ActionTypes.GET_CATEGORIES_TREE] = {};
    }

    if (
      typeof ApiIteratorStore[ActionTypes.GET_CATEGORIES_TREE][buildingId] ===
        "undefined" ||
      forceUpdate
    ) {
      ApiIteratorStore[ActionTypes.GET_CATEGORIES_TREE][buildingId] =
        spinalAPI.createIteratorCall(getAllCategoriesTree, buildingId, context);
    }

    const items = await ApiIteratorStore[ActionTypes.GET_CATEGORIES_TREE][
      buildingId
    ].next();
    return items?.value;
  },

  ////////////////////////////////////////////////////////
  //					TICKET
  ////////////////////////////////////////////////////////

  async [ActionTypes.LOAD_WORKFLOWS](
    { commit, dispatch, state }: AugmentedActionContextAppData,
    { buildingId, config }
  ): Promise<{ [key: string]: INodeItem }> {
    const spinalAPI = SpinalAPI.getInstance();

    if (typeof ApiIteratorStore[ActionTypes.LOAD_WORKFLOWS] === "undefined") {
      ApiIteratorStore[ActionTypes.LOAD_WORKFLOWS] = {};
    }

    // if (
    //   typeof ApiIteratorStore[ActionTypes.LOAD_TICKETS][buildingId] ===
    //   "undefined"
    // ) {
    ApiIteratorStore[ActionTypes.LOAD_WORKFLOWS][buildingId] =
      spinalAPI.createIteratorCall(getWorkflowList);
    // }

    const items = await ApiIteratorStore[ActionTypes.LOAD_WORKFLOWS][
      buildingId
    ].next();
    return items?.value;
  },

  async [ActionTypes.LOAD_PROCESS](
    { commit, dispatch, state }: AugmentedActionContextAppData,
    { buildingId, config, workflowlist }
  ): Promise<{ [key: string]: INodeItem }> {
    const spinalAPI = SpinalAPI.getInstance();
    if (typeof ApiIteratorStore[ActionTypes.LOAD_PROCESS] === "undefined") {
      ApiIteratorStore[ActionTypes.LOAD_PROCESS] = {};
    }

    ApiIteratorStore[ActionTypes.LOAD_PROCESS][buildingId] =
      spinalAPI.createIteratorCall(getProcessList, workflowlist);

    const items = await ApiIteratorStore[ActionTypes.LOAD_PROCESS][
      buildingId
    ].next();
    return items?.value;
  },

  async [ActionTypes.LOAD_TICKETS](
    { commit, dispatch, state }: AugmentedActionContextAppData,
    { buildingId, config }
  ): Promise<{ [key: string]: INodeItem }> {
    const spinalAPI = SpinalAPI.getInstance();

    if (typeof ApiIteratorStore[ActionTypes.LOAD_TICKETS] === "undefined") {
      ApiIteratorStore[ActionTypes.LOAD_TICKETS] = {};
    }

    // if (
    //   typeof ApiIteratorStore[ActionTypes.LOAD_TICKETS][buildingId] ===
    //   "undefined"
    // ) {
    ApiIteratorStore[ActionTypes.LOAD_TICKETS][buildingId] =
      spinalAPI.createIteratorCall(loadTickets, buildingId, config);
    // }

    const items = await ApiIteratorStore[ActionTypes.LOAD_TICKETS][
      buildingId
    ].next();
    return items?.value;
  },

  async [ActionTypes.FILTER_TICKETS](
    { commit, dispatch, state }: AugmentedActionContextAppData,
    { buildingId, dynamicId }
  ) {
    let ticketsToFilter = await dispatch(ActionTypes.LOAD_TICKETS, {
      buildingId,
    });
    console.log(
      "ticketsToFilter eyyyyyyyyyyyyyyyyyyyyyyyyyyyyy",
      ticketsToFilter
    );
    return filterTicketsOnPosition(ticketsToFilter, buildingId, dynamicId);
  },

  async [ActionTypes.ADD_TICKET](
    { commit }: AugmentedActionContextAppData,
    { buildingId, data, file }: { buildingId: string; data: any; file: any[] }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await createTicket(buildingId, data);
      if (!file || file.length === 0) return result;
      file.forEach(async (element) => {
        const file = new FormData();
        file.append("file", element);
        const adddoc = await addTicketDoc(buildingId, result.dynamicId, file);
      });
      return result;
    } catch (error) {
      console.error("Erreur lors de la création du ticket:", error);
      throw error;
    }
  },

  async [ActionTypes.ARCHIVE_TICKET](
    { commit }: AugmentedActionContextAppData,
    {
      buildingId,
      ticketId,
      data,
    }: { buildingId: string; ticketId: any; data: any }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await archiveTicket(buildingId, ticketId, data);

      return result;
    } catch (error) {
      console.error("Erreur lors de la création du ticket:", error);
      throw error;
    }
  },
  async [ActionTypes.MOVE_TO_STEP_TICKET](
    { commit }: AugmentedActionContextAppData,
    {
      buildingId,
      ticketId,
      data,
    }: { buildingId: string; ticketId: any; data: any }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await moveToStepTicket(buildingId, ticketId, data);

      return result;
    } catch (error) {
      console.error("Erreur lors du changement d'etape du ticket:", error);
      throw error;
    }
  },
  async [ActionTypes.ADD_NOTE](
    { commit }: AugmentedActionContextAppData,
    {
      buildingId,
      ticketId,
      data,
    }: { buildingId: string; ticketId: any; data: any }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await addNoteTicket(buildingId, ticketId, data);

      return result;
    } catch (error) {
      console.error("Erreur lors de la création du ticket:", error);
      throw error;
    }
  },
  async [ActionTypes.MODIFY_TICKET](
    { commit }: AugmentedActionContextAppData,
    {
      buildingId,
      ticketId,
      data,
    }: { buildingId: string; ticketId: any; data: any }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await modify_ticket(buildingId, ticketId, data);

      return result;
    } catch (error) {
      console.error("Erreur lors de la création du ticket:", error);
      throw error;
    }
  },
  async [ActionTypes.ADD_DOC](
    { commit }: AugmentedActionContextAppData,
    {
      buildingId,
      ticketId,
      data,
    }: { buildingId: string; ticketId: any; data: any }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      const result = await addTicketDoc(buildingId, ticketId, data);

      return result;
    } catch (error) {
      console.error("Erreur lors de la création du ticket:", error);
      throw error;
    }
  },

  // Viewer////////////////////////////:

  async [ActionTypes.OPEN_VIEWER](
    { commit, dispatch, state }: AugmentedActionContextAppData,
    playload: { onlyThisModel: boolean; config: IConfig; item: any }
  ): Promise<void> {
    try {
      const viewerInfo = playload.config.viewerInfo;
      const body = {
        dynamicId: [playload.item.dynamicId],
        roomRef: viewerInfo.roomRef,
        floorRef: viewerInfo.floorRef,
        equipements: false,
        dbIdsToAdd: [],
      };

      if (viewerInfo.equipments === "all") {
        body.equipements = true;
        body.dbIdsToAdd = [];
      } else if (viewerInfo.equipments === "groupItem") {
        body.equipements = false;
        const map = await dispatch(ActionTypes.GET_GROUPS_ITEMS, {
          config: playload.config,
          buildingId: playload.item.buildingId,
        });
        body.dbIdsToAdd = classifyItemByBimFileId(
          map,
          playload.item.dynamicId,
          playload.item.type
        );
      } else if (viewerInfo.equipments === "ticket") {
        body.equipements = false;
        const map = await dispatch(ActionTypes.FILTER_TICKETS, {
          buildingId: playload.item.buildingId,
          dynamicId: playload.item.dynamicId,
        });
        const eq = map
          .map((t) =>
            t.elementSelected.type === "BIMObject"
              ? t.elementSelected.dynamicId
              : null
          )
          .filter((t) => t);
        const details = await equipment_read_details_multiple([
          ...new Set<any>(eq),
        ]);
        body.dbIdsToAdd = details.reduce((acc, curr) => {
          const found = acc.find((d) => d.bimFileId === curr.bimFileId);
          if (found) {
            if (!found.dbIds.includes(curr.dbid)) found.dbIds.push(curr.dbid);
          } else {
            acc.push({ bimFileId: curr.bimFileId, dbIds: [curr.dbid] });
          }
          return acc;
        }, []);
      }

      await ViewerManager.getInstance().loadInViewer(
        playload.item,
        playload.onlyThisModel,
        body
      );

      if (playload.onlyThisModel) state.viewerStartedList = {};

      commit(MutationTypes.ADD_VIEWER_LOADED, { id: playload.item.dynamicId });
    } catch (error) {
      console.log("errror", error);
    }
  },

  async [ActionTypes.GET_VIEWER_INFO](
    { commit, state }: AugmentedActionContextAppData,
    playload
  ): Promise<IViewInfoItemRes[]> {
    return ViewerManager.getInstance().getViewerInfoMerged(playload);
  },
  [ActionTypes.SELECT_ITEMS]({ commit, dispatch, state }, playload: any) {
    ViewerManager.getInstance().select(playload);
  },

  [ActionTypes.ISOLATE_ITEMS]({ commit, dispatch, state }, playload: any) {
    ViewerManager.getInstance().isolate(playload);
  },

  [ActionTypes.FIT_TO_VIEW_ITEMS]({ commit, dispatch, state }, playload: any) {
    ViewerManager.getInstance().fitToView(playload);
  },

  async [ActionTypes.UNLOAD_MODEL]({ commit, dispatch, state }, playload: any) {
    await ViewerManager.getInstance().unload(playload);

    if (!Array.isArray(playload)) playload = [playload];

    for (const item of playload) {
      commit(MutationTypes.REMOVE_VIEWER_LOADED, { id: item.dynamicId });
    }
  },
  async [ActionTypes.GET_BIM_OBJECT_INFO](
    { commit }: AugmentedActionContextAppData,
    { buildingId, referenceIds }: { buildingId: string; referenceIds: any }
  ): Promise<any> {
    try {
      const result = await postBIMObjectInfo(buildingId, referenceIds);
      return result;
    } catch (error) {
      console.error(
        "Erreur lors de la récupération des objets de référence:",
        error
      );
      throw error;
    }
  },
  async [ActionTypes.GET_STATIC_DETAILS](
    { commit }: AugmentedActionContextAppData,
    { buildingId, referenceIds }: { buildingId: string; referenceIds: number }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      // const result = await spinalAPI.createIteratorCall(getMultipleReferenceObjects, buildingId, referenceIds);
      const result = await getStaticDetails(buildingId, referenceIds);
      return result;
    } catch (error) {
      console.error(
        "Erreur lors de la récupération des objets de référence:",
        error
      );
      throw error;
    }
  },
  async [ActionTypes.GET_STATIC_DETAILS_EQUIPEMENT](
    { commit }: AugmentedActionContextAppData,
    { buildingId, referenceIds }: { buildingId: string; referenceIds: number }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      // const result = await spinalAPI.createIteratorCall(getMultipleReferenceObjects, buildingId, referenceIds);
      const result = await getStaticDetailsEquipement(buildingId, referenceIds);
      return result;
    } catch (error) {
      console.error(
        "Erreur lors de la récupération des objets de référence:",
        error
      );
      throw error;
    }
  },
  async [ActionTypes.GET_REFERENCE_OBJECT_LIST_MULTIPLE](
    { commit }: AugmentedActionContextAppData,
    { buildingId, referenceIds }: { buildingId: string; referenceIds: number[] }
  ): Promise<any> {
    const spinalAPI = SpinalAPI.getInstance();
    try {
      // const result = await spinalAPI.createIteratorCall(getMultipleReferenceObjects, buildingId, referenceIds);
      const result = await getMultipleReferenceObjects(
        buildingId,
        referenceIds
      );
      return result;
    } catch (error) {
      console.error(
        "Erreur lors de la récupération des objets de référence:",
        error
      );
      throw error;
    }
  },

  // [ActionTypes.COLOR_ITEMS](
  //   { commit, dispatch, state },
  //   { items, buildingId }: any
  // ) {
  //   return ViewerManager.getInstance().colorItems(items, buildingId);
  // },
  [ActionTypes.COLOR_ITEMS](context, payload) {
    return ViewerManager.getInstance().colorItems(payload, payload.buildingId);
  },

  [ActionTypes.ADD_SPRITES](
    { commit, dispatch, state },
    { items, buildingId }: any
  ) {
    return ViewerManager.getInstance().addSprites(items, buildingId);
  },

  [ActionTypes.ADD_COMPONENT_AS_SPRITES](
    { commit, dispatch, state },
    { items, buildingId, component }: any
  ) {
    return ViewerManager.getInstance().addComponentAsSprites(
      items,
      buildingId,
      component
    );
  },
  [ActionTypes.ADD_CARD_COMPONENT](
    { commit, dispatch, state },
    { items, buildingId, component }: any
  ) {
    return ViewerManager.getInstance().addCardomponent(
      items,
      buildingId,
      component
    );
  },
  [ActionTypes.REMOVE_CARDS]({ commit, dispatch, state }) {
    return SpriteManager.getInstance().removeCards();
  },

  [ActionTypes.REMOVE_ALL_SPRITES]({ commit, dispatch, state }) {
    return SpriteManager.getInstance().removeSprites();
  },
  [ActionTypes.SELECT_SPRITES](
    { commit, dispatch, state },
    dynamicIds: Array<number>
  ) {
    return SpriteManager.getInstance().selectSprites(dynamicIds);
  },
  async [ActionTypes.RESET_API_ITERATOR_STORE](
    { commit }: AugmentedActionContextAppData,
    { buildingId }: any
  ): Promise<void> {
    // List of ActionTypes keys to check and reset
    const keysToReset = [ActionTypes.ADD_COMPONENT_AS_SPRITES];

    // Iterate through each key and delete the buildingId entry if it exists
    keysToReset.forEach((key) => {
      if (
        typeof ApiIteratorStore[key] !== "undefined" &&
        typeof ApiIteratorStore[key][buildingId] !== "undefined"
      ) {
        delete ApiIteratorStore[key][buildingId];
      }
    });
  },
};
